# add faster implementation for `findall(::AbstractVector{Bool})`