@jakobnissen 2021-09-16
There is already: https://github.com/JuliaLang/julia/pull/42202