@mzgubic 2021-09-16
Oh, nice! Is your approach even faster for the benchmarks above?