@jakobnissen 2021-09-16
Yes - on my branch:
```
julia> a = rand(Bool, 10000);

julia> using BenchmarkTools

julia> @btime findall($a);
  5.085 μs (2 allocations: 38.92 KiB)

julia> function myfindall(a)
                  return (1:length(a))[a]
              end
myfindall (generic function with 1 method)

julia> @btime myfindall($a);
  29.697 μs (2 allocations: 38.92 KiB)
```